+++ app/modules/weekly_report/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.13.0'
+__version__ = '0.14.0'
 
 from .routes import weekly_report_bp
 from .service import WeeklyReportService